refactor: generic.py restructuring
