+++ src/ansys/fluent/core/utils/event_loop.py
@@ -1,0 +1,26 @@
+"""Provides a module for event loop."""
+
+import asyncio
+from functools import partial
+
+loop = asyncio.get_event_loop()
+
+
+def execute_in_event_loop_threadsafe(f):
+    """Decorator to execute function in an event loop from another thread."""
+
+    def cb(*args, **kwargs):
+        par = partial(f, *args, **kwargs)
+        loop.call_soon_threadsafe(par)
+
+    return cb
+
+
+def execute_in_event_loop(f):
+    """Decorator to execute function in an event loop."""
+
+    def cb(*args, **kwargs):
+        par = partial(f, *args, **kwargs)
+        loop.call_soon(par)
+
+    return cb
